backend/requirements.txt
django==3.2
django-grappelli==2.15.1
psycopg2-binary==2.8.6
pillow==8.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
redis==3.5.3
django-cors-headers==3.7.0
channels==3.0.3
channels_redis==3.2.0